2026.03.30 arkweb_144代码蓝黄同步

+++ chromium_ext/content/browser/ohos/content_view_statics_ohos.cc
@@ -26,7 +26,7 @@ void SuspendedProcessWatcherOHOS::PauseWebKitShardTimersFromOHOS() {
       LOG(INFO) << "PauseWebKitShardTimersFromOHOS host is nullptr.";
       continue;
     }
-    if (pause_processes_.insert(host->GetID()).second) {
+    if (pause_processes_.insert(static_cast<int>(host->GetID())).second) {
       host->AddObserver(this);
       host->GetRendererInterface()->SetWebKitSharedTimersSuspended(true);
     }
